% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Algorithm to find the greatest common divisor of two numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{is greater than 0 while x is whole number}

\item{y}{is greater than 0 while y is whole number}
}
\value{
The greatest common divisor is the largest natural number that divides both x and y without a remainder.
}
\description{
The algorithm computes the greatest common divisor of two natural numbers x and y.
}
\examples{
euclidean(x = 1071, y = 462)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
